feature/with_fee_widget: implemented #1

See merge request clients/vexl-android!79

+++ core/src/main/java/cz/cleevio/core/model/FeeValue.kt
@@ -4,5 +4,5 @@ import cz.cleevio.core.widget.FeeButtonSelected
 
 data class FeeValue(
 	val type: FeeButtonSelected,
-	val value: Int
+	val value: Float
 )
